fixed the issue by puttin all package initialisations in the same file
