+++ _src-lp/blocks/header/header.js
@@ -2,7 +2,7 @@ import { getMetadata, decorateIcons2 } from '../../scripts/lib-franklin.js';
 import {
   adobeMcAppendVisitorId, getLocalizedResourceUrl, getDefaultBaseUrl, getDefaultSection, getLocale,
 } from '../../scripts/utils.js';
-import { getDefaultLanguage } from '../../scripts/target.js';
+import { getDefaultLanguage } from '../../scripts/packages.js';
 
 /**
  * extracts the span svg's from the nav html
